Initial commit with first implementations of ati
There are two packages 'ati' and 'atimathsym'. 'ati' implicitly uses
'atimathsym' to use predefined math symbols for tasks.
The package 'ati' defines several environments to define new tasks and
there solutions. The key-value system allows to set arbitrary keys
that do not have to used or defined. At this moment it is only
implemented via 'pgfkeys'. But there should be possibilities to extend
this implementation detail to other packages as well.
For displaying tasks and there solutions the two options 'custom' and
'exsheets' are available. Both of them are used to implement the
abstract task interface.